fix: quantize denoised images to uint8, fix normalize() dtype bug
denoise() already rounds BM3D output to uint8 internally, so the signal
never had more than 8 bits of real precision — but normalize()'s
percentile-clip rescale was silently upcasting to float64 (np.clip
against np.percentile's float64 bounds promotes under NumPy 2.x), and
pipeline.py saved that float64 array directly. Result: denoised .npy
files were 4x larger than the data actually justified, for zero
precision gain.

- normalize(): explicit .astype(np.float32) at the end, matching its
  own docstring
- pipeline.py: preprocess() now re-quantizes the normalized [0, 1]
  result to uint8 before saving
- dataset.py: OCTDataset.__getitem__ expands uint8 -> float32 [0, 1] at
  load time, so every consumer keeps the same [0, 1] float contract
- visualize_boundaries.py: same expansion, since it loads .npy directly
  rather than through OCTDataset

+++ src/s2_preprocessing/normalize.py
@@ -19,4 +19,4 @@ def normalize(bscan: np.ndarray, lo_pct: float = 1, hi_pct: float = 99) -> np.nd
     clipped = np.clip(img_float, lo, hi)
     normalizedScan = (clipped - lo) / (hi - lo)
 
-    return normalizedScan
+    return normalizedScan.astype(np.float32)
